refactor(bert): move BertArgs and BertVariant into bert_config
`src/models/bert.rs` had grown to 562 lines, past the repository's 500-line file guideline. Config resolution is the natural seam: `BertVariant`, the serde shape and `BertArgs` share no code with the block stack below them.

`bert_config` stays a private module and `bert` re-exports both types, so `crate::models::bert::BertArgs` remains the single public path and no caller, test or doc link changes. The encoder module is now 387 lines and the config module 199.

The three BERT test modules and the real-checkpoint gates were re-run after the split: 27 passed, 0 failed, with every gate number identical to the pre-split run.
